Revert "chore: improve dependency injection (#9)"
This reverts commit b9743afb8f0c93439d5fec03981261c696c79151.
