--- pkg/engagement/usecases/feed/config_test.go
@@ -1,236 +1,0 @@
-package feed_test
-
-import (
-	"encoding/json"
-	"math/rand"
-	"os"
-	"testing"
-	"time"
-
-	"github.com/savannahghi/feedlib"
-	"github.com/segmentio/ksuid"
-	"github.com/stretchr/testify/assert"
-)
-
-func TestMain(m *testing.M) {
-	os.Setenv("ROOT_COLLECTION_SUFFIX", "staging")
-	os.Setenv("ENVIRONMENT", "staging")
-	os.Exit(m.Run())
-}
-
-func getTestMessage() feedlib.Message {
-	return feedlib.Message{
-		ID:             ksuid.New().String(),
-		SequenceNumber: getTestSequenceNumber(),
-		Text:           ksuid.New().String(),
-		ReplyTo:        ksuid.New().String(),
-		PostedByUID:    ksuid.New().String(),
-		PostedByName:   ksuid.New().String(),
-		Timestamp:      time.Now(),
-	}
-}
-
-func getInvalidTestMessage() feedlib.Message {
-	return feedlib.Message{
-		ID:             "",
-		SequenceNumber: 0,
-		Text:           ksuid.New().String(),
-		ReplyTo:        ksuid.New().String(),
-		PostedByUID:    ksuid.New().String(),
-		PostedByName:   ksuid.New().String(),
-		Timestamp:      time.Now(),
-	}
-}
-
-func getTestSequenceNumber() int {
-	return rand.Intn(IntMax)
-}
-
-func testItem() *feedlib.Item {
-	return &feedlib.Item{
-		ID:             ksuid.New().String(),
-		SequenceNumber: getTestSequenceNumber(),
-		Expiry:         getTextExpiry(),
-		Persistent:     true,
-		Status:         feedlib.StatusPending,
-		Visibility:     feedlib.VisibilityShow,
-		Icon: feedlib.GetPNGImageLink(
-			feedlib.LogoURL, "title", "description", feedlib.BlankImageURL),
-		Author:    ksuid.New().String(),
-		Tagline:   ksuid.New().String(),
-		Label:     ksuid.New().String(),
-		Timestamp: time.Now(),
-		Summary:   ksuid.New().String(),
-		Text:      ksuid.New().String(),
-		TextType:  feedlib.TextTypeMarkdown,
-		Links: []feedlib.Link{
-			feedlib.GetPNGImageLink(
-				feedlib.LogoURL, "title", "description", feedlib.BlankImageURL),
-		},
-		Actions: []feedlib.Action{
-			getTestAction(),
-		},
-		Conversations: []feedlib.Message{
-			getTestMessage(),
-		},
-		Users: []string{
-			ksuid.New().String(),
-		},
-		Groups: []string{
-			ksuid.New().String(),
-		},
-		NotificationChannels: []feedlib.Channel{
-			feedlib.ChannelEmail,
-			feedlib.ChannelFcm,
-			feedlib.ChannelSms,
-			feedlib.ChannelWhatsapp,
-		},
-	}
-}
-
-func getTextExpiry() time.Time {
-	return time.Now().Add(time.Hour * 24000)
-}
-
-func getTestEvent() feedlib.Event {
-	return feedlib.Event{
-		ID:   ksuid.New().String(),
-		Name: "TEST_EVENT",
-		Context: feedlib.Context{
-			UserID:         ksuid.New().String(),
-			Flavour:        feedlib.FlavourConsumer,
-			OrganizationID: ksuid.New().String(),
-			LocationID:     ksuid.New().String(),
-			Timestamp:      time.Now(),
-		},
-	}
-}
-
-func getInvalidTestEvent() feedlib.Event {
-	return feedlib.Event{
-		ID:   "",
-		Name: "TEST_EVENT",
-		Context: feedlib.Context{
-			UserID:         "",
-			Flavour:        "",
-			OrganizationID: ksuid.New().String(),
-			LocationID:     ksuid.New().String(),
-			Timestamp:      time.Now(),
-		},
-	}
-}
-
-func getTestAction() feedlib.Action {
-	return feedlib.Action{
-		ID:             ksuid.New().String(),
-		SequenceNumber: getTestSequenceNumber(),
-		Name:           "TEST_ACTION",
-		Icon: feedlib.GetPNGImageLink(
-			feedlib.LogoURL, "title", "description", feedlib.BlankImageURL),
-		ActionType: feedlib.ActionTypePrimary,
-		Handling:   feedlib.HandlingFullPage,
-	}
-}
-
-func testNudge() *feedlib.Nudge {
-	return &feedlib.Nudge{
-		ID:             ksuid.New().String(),
-		SequenceNumber: getTestSequenceNumber(),
-		Status:         feedlib.StatusPending,
-		Visibility:     feedlib.VisibilityShow,
-		Title:          ksuid.New().String(),
-		Links: []feedlib.Link{
-			feedlib.GetPNGImageLink(feedlib.LogoURL, "title", "description", feedlib.BlankImageURL),
-		},
-		Text: ksuid.New().String(),
-		Actions: []feedlib.Action{
-			getTestAction(),
-		},
-		Users: []string{
-			ksuid.New().String(),
-		},
-		Groups: []string{
-			ksuid.New().String(),
-		},
-		NotificationChannels: []feedlib.Channel{
-			feedlib.ChannelEmail,
-			feedlib.ChannelFcm,
-			feedlib.ChannelSms,
-			feedlib.ChannelWhatsapp,
-		},
-	}
-}
-
-func getEmptyJson(t *testing.T) []byte {
-	emptyJSONBytes, err := json.Marshal(map[string]string{})
-	assert.Nil(t, err)
-	assert.NotNil(t, emptyJSONBytes)
-	return emptyJSONBytes
-}
-
-func getTestItem() feedlib.Item {
-	return feedlib.Item{
-		ID:             "item-1",
-		SequenceNumber: 1,
-		Expiry:         time.Now(),
-		Persistent:     true,
-		Status:         feedlib.StatusPending,
-		Visibility:     feedlib.VisibilityShow,
-		Icon:           feedlib.GetPNGImageLink(feedlib.LogoURL, "title", "description", feedlib.BlankImageURL),
-		Author:         "Bot 1",
-		Tagline:        "Bot speaks...",
-		Label:          "DRUGS",
-		Timestamp:      time.Now(),
-		Summary:        "I am a bot...",
-		Text:           "This bot can speak",
-		TextType:       feedlib.TextTypePlain,
-		Links: []feedlib.Link{
-			feedlib.GetYoutubeVideoLink(sampleVideoURL, "title", "description", feedlib.BlankImageURL),
-		},
-		Actions: []feedlib.Action{
-			{
-				ID:             ksuid.New().String(),
-				SequenceNumber: 1,
-				Name:           "ACTION_NAME",
-				Icon:           feedlib.GetPNGImageLink(feedlib.LogoURL, "title", "description", feedlib.BlankImageURL),
-				ActionType:     feedlib.ActionTypeSecondary,
-				Handling:       feedlib.HandlingFullPage,
-				AllowAnonymous: false,
-			},
-			{
-				ID:             "action-1",
-				SequenceNumber: 1,
-				Name:           "First action",
-				Icon:           feedlib.GetPNGImageLink(feedlib.LogoURL, "title", "description", feedlib.BlankImageURL),
-				ActionType:     feedlib.ActionTypePrimary,
-				Handling:       feedlib.HandlingInline,
-				AllowAnonymous: true,
-			},
-		},
-		Conversations: []feedlib.Message{
-			{
-				ID:             "msg-2",
-				SequenceNumber: 1,
-				Text:           "hii ni reply",
-				ReplyTo:        "msg-1",
-				PostedByName:   ksuid.New().String(),
-				PostedByUID:    ksuid.New().String(),
-				Timestamp:      time.Now(),
-			},
-		},
-		Users: []string{
-			"user-1",
-			"user-2",
-		},
-		Groups: []string{
-			"group-1",
-			"group-2",
-		},
-		NotificationChannels: []feedlib.Channel{
-			feedlib.ChannelFcm,
-			feedlib.ChannelEmail,
-			feedlib.ChannelSms,
-			feedlib.ChannelWhatsapp,
-		},
-	}
-}
