+++ pkg/engagement/usecases/feed/feed_test.go
@@ -3,6 +3,7 @@ package feed_test
 import (
 	"context"
 	"encoding/json"
+	"math/rand"
 	"reflect"
 	"testing"
 	"time"
@@ -27,6 +28,127 @@ func InitializeTestNewFeed(ctx context.Context) (*feed.UseCaseImpl, infrastructu
 	infra := infrastructure.NewInteractor()
 	feed := feed.NewFeed(infra)
 	return feed, infra, nil
+}
+
+// var fakeEngagement mockEngagement.FakeEngagementRepository
+// var fakeOnboarding mockOnboarding.FakeServiceOnboarding
+// var fakeMessaging mockMessaging.FakeServiceMessaging
+// var fakeFCM mockFCM.FakeServiceFcm
+
+// // InitializeFakeEngagementInteractor represents a fake engagement interactor
+// func InitializeFakeEngagementInteractor() (*interactor.Interactor, error) {
+// 	var r repository.Repository = &fakeEngagement
+// 	var onboardingSvc onboarding.ProfileService = &fakeOnboarding
+// 	var messagingSvc messaging.NotificationService = &fakeMessaging
+// 	var fcmSvc fcm.PushService = &fakeFCM
+
+// 	feed := usecases.NewFeed(r, messagingSvc)
+// 	fcm := fcm.NewService(r, onboardingSvc)
+// 	mail := mail.NewService(r)
+// 	notification := usecases.NewNotification(r, fcmSvc, onboardingSvc, fcm, mail)
+// 	uploads := uploads.NewUploadsService()
+
+// 	library := library.NewLibraryService(onboardingSvc)
+
+// 	sms := sms.NewService(r, messagingSvc)
+// 	whatsapp := whatsapp.NewService()
+// 	twilio := twilio.NewService(sms, r)
+// 	otp := otp.NewService(whatsapp, mail, sms, twilio)
+// 	surveys := surveys.NewService(r)
+
+// 	i, err := interactor.NewEngagementInteractor(
+// 		feed,
+// 		notification,
+// 		uploads,
+// 		library,
+// 		sms,
+// 		*mail,
+// 		whatsapp,
+// 		otp,
+// 		twilio,
+// 		fcm,
+// 		surveys,
+// 		onboardingSvc,
+// 	)
+
+// 	if err != nil {
+// 		return nil, fmt.Errorf("can't instantiate service : %w", err)
+// 	}
+// 	return i, nil
+// }
+
+func getEmptyJson(t *testing.T) []byte {
+	emptyJSONBytes, err := json.Marshal(map[string]string{})
+	assert.Nil(t, err)
+	assert.NotNil(t, emptyJSONBytes)
+	return emptyJSONBytes
+}
+
+func getTestItem() feedlib.Item {
+	return feedlib.Item{
+		ID:             "item-1",
+		SequenceNumber: 1,
+		Expiry:         time.Now(),
+		Persistent:     true,
+		Status:         feedlib.StatusPending,
+		Visibility:     feedlib.VisibilityShow,
+		Icon:           feedlib.GetPNGImageLink(feedlib.LogoURL, "title", "description", feedlib.BlankImageURL),
+		Author:         "Bot 1",
+		Tagline:        "Bot speaks...",
+		Label:          "DRUGS",
+		Timestamp:      time.Now(),
+		Summary:        "I am a bot...",
+		Text:           "This bot can speak",
+		TextType:       feedlib.TextTypePlain,
+		Links: []feedlib.Link{
+			feedlib.GetYoutubeVideoLink(sampleVideoURL, "title", "description", feedlib.BlankImageURL),
+		},
+		Actions: []feedlib.Action{
+			{
+				ID:             ksuid.New().String(),
+				SequenceNumber: 1,
+				Name:           "ACTION_NAME",
+				Icon:           feedlib.GetPNGImageLink(feedlib.LogoURL, "title", "description", feedlib.BlankImageURL),
+				ActionType:     feedlib.ActionTypeSecondary,
+				Handling:       feedlib.HandlingFullPage,
+				AllowAnonymous: false,
+			},
+			{
+				ID:             "action-1",
+				SequenceNumber: 1,
+				Name:           "First action",
+				Icon:           feedlib.GetPNGImageLink(feedlib.LogoURL, "title", "description", feedlib.BlankImageURL),
+				ActionType:     feedlib.ActionTypePrimary,
+				Handling:       feedlib.HandlingInline,
+				AllowAnonymous: true,
+			},
+		},
+		Conversations: []feedlib.Message{
+			{
+				ID:             "msg-2",
+				SequenceNumber: 1,
+				Text:           "hii ni reply",
+				ReplyTo:        "msg-1",
+				PostedByName:   ksuid.New().String(),
+				PostedByUID:    ksuid.New().String(),
+				Timestamp:      time.Now(),
+			},
+		},
+		Users: []string{
+			"user-1",
+			"user-2",
+		},
+		Groups: []string{
+			"group-1",
+			"group-2",
+		},
+		NotificationChannels: []feedlib.Channel{
+			feedlib.ChannelFcm,
+			feedlib.ChannelEmail,
+			feedlib.ChannelSms,
+			feedlib.ChannelWhatsapp,
+		},
+	}
 }
 
 func TestMessage_ValidateAndUnmarshal(t *testing.T) {
@@ -2960,6 +3082,149 @@ func TestFeed_ProcessEvent(t *testing.T) {
 				t.Errorf("Feed.ProcessEvent() error = %v, wantErr %v", err, tt.wantErr)
 			}
 		})
+	}
+}
+
+func getTestMessage() feedlib.Message {
+	return feedlib.Message{
+		ID:             ksuid.New().String(),
+		SequenceNumber: getTestSequenceNumber(),
+		Text:           ksuid.New().String(),
+		ReplyTo:        ksuid.New().String(),
+		PostedByUID:    ksuid.New().String(),
+		PostedByName:   ksuid.New().String(),
+		Timestamp:      time.Now(),
+	}
+}
+
+func getInvalidTestMessage() feedlib.Message {
+	return feedlib.Message{
+		ID:             "",
+		SequenceNumber: 0,
+		Text:           ksuid.New().String(),
+		ReplyTo:        ksuid.New().String(),
+		PostedByUID:    ksuid.New().String(),
+		PostedByName:   ksuid.New().String(),
+		Timestamp:      time.Now(),
+	}
+}
+
+func getTestSequenceNumber() int {
+	return rand.Intn(IntMax)
+}
+
+func testItem() *feedlib.Item {
+	return &feedlib.Item{
+		ID:             ksuid.New().String(),
+		SequenceNumber: getTestSequenceNumber(),
+		Expiry:         getTextExpiry(),
+		Persistent:     true,
+		Status:         feedlib.StatusPending,
+		Visibility:     feedlib.VisibilityShow,
+		Icon: feedlib.GetPNGImageLink(
+			feedlib.LogoURL, "title", "description", feedlib.BlankImageURL),
+		Author:    ksuid.New().String(),
+		Tagline:   ksuid.New().String(),
+		Label:     ksuid.New().String(),
+		Timestamp: time.Now(),
+		Summary:   ksuid.New().String(),
+		Text:      ksuid.New().String(),
+		TextType:  feedlib.TextTypeMarkdown,
+		Links: []feedlib.Link{
+			feedlib.GetPNGImageLink(
+				feedlib.LogoURL, "title", "description", feedlib.BlankImageURL),
+		},
+		Actions: []feedlib.Action{
+			getTestAction(),
+		},
+		Conversations: []feedlib.Message{
+			getTestMessage(),
+		},
+		Users: []string{
+			ksuid.New().String(),
+		},
+		Groups: []string{
+			ksuid.New().String(),
+		},
+		NotificationChannels: []feedlib.Channel{
+			feedlib.ChannelEmail,
+			feedlib.ChannelFcm,
+			feedlib.ChannelSms,
+			feedlib.ChannelWhatsapp,
+		},
+	}
+}
+
+func getTextExpiry() time.Time {
+	return time.Now().Add(time.Hour * 24000)
+}
+
+func getTestEvent() feedlib.Event {
+	return feedlib.Event{
+		ID:   ksuid.New().String(),
+		Name: "TEST_EVENT",
+		Context: feedlib.Context{
+			UserID:         ksuid.New().String(),
+			Flavour:        feedlib.FlavourConsumer,
+			OrganizationID: ksuid.New().String(),
+			LocationID:     ksuid.New().String(),
+			Timestamp:      time.Now(),
+		},
+	}
+}
+
+func getInvalidTestEvent() feedlib.Event {
+	return feedlib.Event{
+		ID:   "",
+		Name: "TEST_EVENT",
+		Context: feedlib.Context{
+			UserID:         "",
+			Flavour:        "",
+			OrganizationID: ksuid.New().String(),
+			LocationID:     ksuid.New().String(),
+			Timestamp:      time.Now(),
+		},
+	}
+}
+
+func getTestAction() feedlib.Action {
+	return feedlib.Action{
+		ID:             ksuid.New().String(),
+		SequenceNumber: getTestSequenceNumber(),
+		Name:           "TEST_ACTION",
+		Icon: feedlib.GetPNGImageLink(
+			feedlib.LogoURL, "title", "description", feedlib.BlankImageURL),
+		ActionType: feedlib.ActionTypePrimary,
+		Handling:   feedlib.HandlingFullPage,
+	}
+}
+
+func testNudge() *feedlib.Nudge {
+	return &feedlib.Nudge{
+		ID:             ksuid.New().String(),
+		SequenceNumber: getTestSequenceNumber(),
+		Status:         feedlib.StatusPending,
+		Visibility:     feedlib.VisibilityShow,
+		Title:          ksuid.New().String(),
+		Links: []feedlib.Link{
+			feedlib.GetPNGImageLink(feedlib.LogoURL, "title", "description", feedlib.BlankImageURL),
+		},
+		Text: ksuid.New().String(),
+		Actions: []feedlib.Action{
+			getTestAction(),
+		},
+		Users: []string{
+			ksuid.New().String(),
+		},
+		Groups: []string{
+			ksuid.New().String(),
+		},
+		NotificationChannels: []feedlib.Channel{
+			feedlib.ChannelEmail,
+			feedlib.ChannelFcm,
+			feedlib.ChannelSms,
+			feedlib.ChannelWhatsapp,
+		},
 	}
 }
